Remove ConstraintSolver and refactor related code
ConstraintSolver functionality has been replaced with a more streamlined Solver implementation, and its files have been deleted. References to ConstraintSolver were refactored in DynamicSystem, simplifying constraint management. Additional improvements include changes to gravity force generators and adjustments to body initialization in Simulation for testing purposes.

+++ include/core/types.h
@@ -8,7 +8,7 @@
 
 namespace Neutron {
 
-using UniqueID = std::uint8_t;
+using UniqueID = std::uint32_t;
 
 // Common vector/matrix types
 using Vector3d = Eigen::Vector3d;
